+++ src/host/main.cpp
@@ -4,7 +4,6 @@
 #include "ccf/ds/logger.h"
 #include "ccf/ds/unit_strings.h"
 #include "ccf/ds/x509_time_fmt.h"
-#include "ccf/pal/attestation.h"
 #include "ccf/pal/platform.h"
 #include "ccf/version.h"
 #include "config_schema.h"
@@ -23,6 +22,7 @@
 #include "lfs_file_handler.h"
 #include "load_monitor.h"
 #include "node_connections.h"
+#include "pal/attestation.h"
 #include "process_launcher.h"
 #include "rpc_connections.h"
 #include "sig_term.h"
